python/requirements.txt
psycopg2-binary>=2.9.3
python-dotenv>=0.19.2
Telethon>=1.24.0
APScheduler~=3.9.1
SQLAlchemy~=1.4.40